Make calc functions and basic html for calculator
Did the first 4 steps of the assignment (forgot to commit in between...):
1 make functions for the 4 basic types of calculation (+ - * /)
2 make three variables for numbers (two variables) and operator.
3 make a function (operate) to get a calculation when you have two numbers
and an operator as input. Use switch-method to choose the right operator.
4 create the html with buttons and input screen, and styled it with css.

After this I researched how to make functions to get the display filled when you click the buttons.
I found a complete calculator (html, css, javascript) on https://www.geeksforgeeks.org/javascript/javascript-calculator/
so I'm trying to understand that one.